feat: added typedoc and vitest
Setup and configured TypeDoc and Vitest. Moved original documentation files to a project directory so that TypeDoc does not overwrite existing data. Will use TSDoc and JSDoc comments with TypeDoc to produce rough documentation. Plan to refactor these documents so that it will be clear documentation in a framework like Docusaurs or MKDocs